+++ src/pages/Home/ArticleCard.module.less
@@ -7,10 +7,11 @@
     background-color: rgb(236, 236, 236);
     // margin: 30px 10px 30px 10px;
     position: relative;
+    box-shadow: 0 3px 10px rgb(0 0 0 / 0.2);
 }
 
 .articletitle {
-    margin: 5px;
+    padding: 10px 5px;
     border-bottom: rgb(151, 151, 151) solid 1px;
     font-size: 20px;
     font-weight: bold;
